requirements.txt
anyio==4.6.2.post1
cairocffi==1.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloup==3.0.5
colorama==0.4.6
colour==0.1.5
commonmark==0.9.1
decorator==5.1.1
glcontext==3.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
isosurfaces==0.1.2
manim==0.18.1
ManimPango==0.6.0
mapbox_earcut==1.0.2
markdown-it-py==3.0.0
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==2.4.6
multipledispatch==1.0.0
networkx==3.4.2
numpy==1.26.4
pangocairocffi==0.4.0
pangocffi==0.8.0
pillow==11.0.0
progressbar==2.5
pycairo==1.27.0
pycparser==2.22
pydub==0.25.1
pyglet==2.0.18
Pygments==2.18.0
pyrr==0.10.3
requests==2.32.3
rich==13.9.4
scipy==1.14.1
screeninfo==0.8.1
skia-pathops==0.8.0.post2
sniffio==1.3.1
srt==3.5.3
svgelements==1.9.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==6.0.0
